fixed tons and tons of warnings cleaned up macros a little
git-svn-id: https://bugengine.googlecode.com/svn/trunk@53 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/graphics/api/graphics/renderer/gpubuffer.hh
@@ -40,7 +40,7 @@ protected:
 public:
     virtual ~GpuBuffer();
 
-    virtual void* map(GpuMapFlags flags, size_t byteCount, size_t byteOffset) = 0;
+    virtual void* map(GpuMapFlags flags, u32 byteCount, u32 byteOffset) = 0;
     virtual void  unmap() = 0;
 };
